@@ -1,13 +1,14 @@
 package com.yohannes.app.dev.newsapp;
 
 import androidx.appcompat.app.AppCompatActivity;
+
 import android.os.Bundle;
 
-public class CreateActivity extends AppCompatActivity {
+public class NoConnection extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_create);
+        setContentView(R.layout.activity_no_connection);
     }
 }
