@@ -1,16 +1,13 @@
 package com.yohannes.app.dev.newsapp;
 
-import android.app.Activity;
 import android.content.Intent;
-import android.media.Image;
-import android.net.Uri;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
-import android.support.v7.widget.Toolbar;
 import android.view.View;
 import android.widget.ImageView;
-import android.widget.ListView;
 import android.widget.TextView;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.Toolbar;
 
 import com.squareup.picasso.Picasso;
 import com.yohannes.app.dev.newsapp.models.SettingsListAdapter;
@@ -31,7 +28,7 @@ public class Settings extends AppCompatActivity {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_settings);
 
-        Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar_settings);
+        Toolbar toolbar = findViewById(R.id.toolbar_settings);
         setSupportActionBar(toolbar);
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
         getSupportActionBar().setDisplayShowHomeEnabled(true);
@@ -43,15 +40,15 @@ public class Settings extends AppCompatActivity {
         settingsValue = new String[]{loggedInUser.getUsername(), loggedInUser.getPassword(), loggedInUser.getName(), loggedInUser.getFname(), String.valueOf(loggedInUser.getPhonenum()), loggedInUser.getBio()};
 
         SettingsListAdapter settingsListAdapter = new SettingsListAdapter(getApplicationContext(), settingsTitles, settingsValue);
-        userAvatarimage = (ImageView) findViewById(R.id.settingsAvatarView);
+        userAvatarimage = findViewById(R.id.settingsAvatarView);
         //ListView settingsListView = (ListView) findViewById(R.id.settings_List);
 
-        TextView userName = (TextView) findViewById(R.id.settings_username_text);
-        TextView password = (TextView) findViewById(R.id.settings_password_text);
-        TextView name = (TextView) findViewById(R.id.settings_name_text);
-        TextView fname = (TextView) findViewById(R.id.settings_fname_text);
-        TextView phonenum = (TextView) findViewById(R.id.settings_phonenum_text);
-        TextView bio = (TextView) findViewById(R.id.settings_bio_text);
+        TextView userName = findViewById(R.id.settings_username_text);
+        TextView password = findViewById(R.id.settings_password_text);
+        TextView name = findViewById(R.id.settings_name_text);
+        TextView fname = findViewById(R.id.settings_fname_text);
+        TextView phonenum = findViewById(R.id.settings_phonenum_text);
+        TextView bio = findViewById(R.id.settings_bio_text);
 
 
         Picasso.with(Settings.this).load(loggedInUser.getAvatar_link()).transform(new CircleAvatar()).into(userAvatarimage);
