@@ -1,15 +1,14 @@
 package com.yohannes.app.dev.newsapp;
 
-import android.app.ProgressDialog;
 import android.content.Context;
 import android.content.Intent;
 import android.os.AsyncTask;
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
 import android.util.Log;
 import android.widget.Toast;
 
-/*import com.google.android.gms.security.ProviderInstaller;*/
 import com.yohannes.app.dev.newsapp.models.News;
 import com.yohannes.app.dev.newsapp.models.User;
 import com.yohannes.app.dev.newsapp.util.DbManager;
@@ -19,18 +18,10 @@ import org.json.JSONArray;
 import org.json.JSONException;
 import org.json.JSONObject;
 
-import java.io.BufferedInputStream;
 import java.io.BufferedReader;
-import java.io.IOException;
-import java.io.InputStream;
 import java.io.InputStreamReader;
-import java.io.OutputStream;
 import java.net.HttpURLConnection;
 import java.net.URL;
-import java.net.URLEncoder;
-import java.util.Arrays;
-
-import javax.net.ssl.HttpsURLConnection;
 
 public class SplashScreen extends AppCompatActivity {
 
